+++ src/Components/ExpressCheckoutUtil.php
@@ -1,0 +1,326 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PaynlPayment\Shopware6\Components;
+
+use Exception;
+use Paynl\Transaction;
+use PaynlPayment\Shopware6\Repository\Country\CountryRepositoryInterface;
+use PaynlPayment\Shopware6\Repository\PaymentMethod\PaymentMethodRepository;
+use PaynlPayment\Shopware6\Repository\Product\ProductRepositoryInterface;
+use PaynlPayment\Shopware6\Repository\SalesChannel\SalesChannelRepositoryInterface;
+use PaynlPayment\Shopware6\Repository\Salutation\SalutationRepositoryInterface;
+use PaynlPayment\Shopware6\Service\Cart\CartBackupService;
+use PaynlPayment\Shopware6\Service\CartServiceInterface;
+use PaynlPayment\Shopware6\Service\CustomerService;
+use PaynlPayment\Shopware6\Service\OrderService;
+use PaynlPayment\Shopware6\ValueObjects\PAY\Order\Amount;
+use PaynlPayment\Shopware6\ValueObjects\PAY\Order\Product;
+use RuntimeException;
+use Shopware\Core\Checkout\Customer\Aggregate\CustomerAddress\CustomerAddressEntity;
+use Shopware\Core\Checkout\Customer\CustomerEntity;
+use Shopware\Core\Checkout\Order\Aggregate\OrderLineItem\OrderLineItemCollection;
+use Shopware\Core\Checkout\Order\Aggregate\OrderLineItem\OrderLineItemEntity;
+use Shopware\Core\Checkout\Order\OrderEntity;
+use Shopware\Core\Framework\Context;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\Framework\Uuid\Uuid;
+use Shopware\Core\Framework\Validation\DataBag\DataBag;
+use Shopware\Core\System\Country\CountryEntity;
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Shopware\Core\System\Salutation\SalutationEntity;
+use Throwable;
+
+class ExpressCheckoutUtil
+{
+    private const ADDRESS_KEYS = [
+        'firstName',
+        'lastName',
+        'street',
+        'zipcode',
+        'countryId',
+        'city',
+        'phoneNumber',
+        'additionalAddressLine1',
+    ];
+
+    /** @var CustomerService */
+    private $customerService;
+
+    /** @var CartServiceInterface */
+    private $cartService;
+
+    /** @var CartBackupService */
+    private $cartBackupService;
+
+    /** @var OrderService */
+    private $orderService;
+
+    /** @var CountryRepositoryInterface */
+    private $countryRepository;
+
+    /** @var PaymentMethodRepository */
+    private $repoPaymentMethods;
+
+    /** @var ProductRepositoryInterface */
+    private $productRepository;
+
+    /** @var SalutationRepositoryInterface */
+    private $salutationRepository;
+
+    /** @var SalesChannelRepositoryInterface */
+    private $salesChannelRepository;
+
+    public function __construct(CustomerService $customerService, CartServiceInterface $cartService, CartBackupService $cartBackupService, OrderService $orderService, CountryRepositoryInterface $countryRepository, PaymentMethodRepository $repoPaymentMethods, ProductRepositoryInterface $productRepository, SalutationRepositoryInterface $salutationRepository, SalesChannelRepositoryInterface $salesChannelRepository)
+    {
+        $this->customerService = $customerService;
+        $this->cartService = $cartService;
+        $this->cartBackupService = $cartBackupService;
+        $this->orderService = $orderService;
+        $this->countryRepository = $countryRepository;
+        $this->repoPaymentMethods = $repoPaymentMethods;
+        $this->productRepository = $productRepository;
+        $this->salutationRepository = $salutationRepository;
+        $this->salesChannelRepository = $salesChannelRepository;
+    }
+
+    public function getActiveIdealID(SalesChannelContext $context): string
+    {
+        return $this->repoPaymentMethods->getActiveIdealID($context->getContext());
+    }
+
+    public function getActivePayPalID(SalesChannelContext $context): string
+    {
+        return $this->repoPaymentMethods->getActivePayPalID($context->getContext());
+    }
+
+    public function prepareCustomer(
+        string $firstname,
+        string $lastname,
+        string $email,
+        string $street,
+        string $zipcode,
+        string $city,
+        string $paymentMethodId,
+        SalesChannelContext $context
+    ): SalesChannelContext {
+        $this->cartBackupService->clearBackup($context);
+
+        if (!$this->customerService->isCustomerLoggedIn($context)) {
+            $countryCode = $this->getSalesChannelCountryIso($context);
+
+            $customer = $this->customerService->createPaymentExpressCustomer(
+                $firstname,
+                $lastname,
+                $email,
+                '',
+                $street,
+                $zipcode,
+                $city,
+                $countryCode,
+                $paymentMethodId,
+                $context
+            );
+
+            if (!$customer instanceof CustomerEntity) {
+                throw new Exception('Error when creating customer!');
+            }
+
+            $this->customerService->customerLogin($customer, $context);
+        }
+
+        // update our payment method to use PayPal Express for our cart
+        return $this->cartService->updatePaymentMethod($context, $paymentMethodId);
+    }
+
+    public function createOrder(SalesChannelContext $context): OrderEntity
+    {
+        $data = new DataBag();
+
+        $data->add(['tos' => true]);
+
+        return $this->orderService->createOrder($data, $context);
+    }
+
+    public function isNotCompletedOrder(string $orderId, Context $context)
+    {
+        try {
+            $order = $this->orderService->getOrder($orderId, $context);
+            $billingAddress = $order->getBillingAddress();
+
+            if (!$billingAddress) {
+                return true;
+            }
+
+            return in_array('Temp', [
+                $billingAddress->getFirstName(),
+                $billingAddress->getLastName(),
+            ]);
+        } catch (Throwable $exception) {
+            return true;
+        }
+    }
+
+    public function getOrderProducts(OrderEntity $order, SalesChannelContext $salesChannelContext): array
+    {
+        $currency = $salesChannelContext->getCurrency()->getIsoCode();
+        $context = $salesChannelContext->getContext();
+
+        /** @var OrderLineItemCollection $orderLineItems*/
+        $orderLineItems = $order->getLineItems();
+        $productsItems = $orderLineItems->filterByProperty('type', 'product');
+        $productsIds = [];
+
+        foreach ($productsItems as $product) {
+            $productsIds[] = $product->getReferencedId();
+        }
+
+        $criteria = new Criteria();
+        $criteria->addFilter(new EqualsAnyFilter('product.id', $productsIds));
+        $entities = $this->productRepository->search($criteria, $context);
+        $elements = $entities->getElements();
+
+        /** @var OrderLineItemEntity $item */
+        foreach ($productsItems as $item) {
+            $vatPercentage = 0;
+            if ($item->getPrice()->getCalculatedTaxes()->first() !== null) {
+                $vatPercentage = $item->getPrice()->getCalculatedTaxes()->first()->getTaxRate();
+            }
+
+            $products[] = new Product(
+                new Amount(
+                    (int) round($item->getUnitPrice() * 100),
+                    $currency
+                ),
+                (string) $elements[$item->getReferencedId()]->get('autoIncrement'),
+                $item->getLabel(),
+                Transaction::PRODUCT_TYPE_ARTICLE,
+                $item->getPrice()->getQuantity(),
+                $vatPercentage
+            );
+        }
+
+        $surchargeItems = $orderLineItems->filterByProperty('type', 'payment_surcharge');
+        /** @var OrderLineItemEntity $item */
+        foreach ($surchargeItems as $item) {
+            $vatPercentage = 0;
+            if ($item->getPrice()->getCalculatedTaxes()->first() !== null) {
+                $vatPercentage = $item->getPrice()->getCalculatedTaxes()->first()->getTaxRate();
+            }
+
+            $products[] = new Product(
+                new Amount(
+                    (int) round($item->getUnitPrice() * 100),
+                    $currency
+                ),
+                'payment',
+                $item->getLabel(),
+                Transaction::PRODUCT_TYPE_PAYMENT,
+                $item->getPrice()->getQuantity(),
+                $vatPercentage
+            );
+        }
+
+        $products[] = new Product(
+            new Amount(
+                (int) round($order->getShippingTotal() * 100),
+                $currency
+            ),
+            'shipping',
+            'Shipping',
+            Transaction::PRODUCT_TYPE_SHIPPING,
+            1,
+            $order->getShippingCosts()->getCalculatedTaxes()->getAmount()
+        );
+
+        return $products;
+    }
+
+    public function getCountryIdByCode(string $code, Context $context): ?string
+    {
+        $criteria = new Criteria();
+        $criteria->addFilter(
+            new EqualsFilter('iso', $code)
+        );
+
+        /** @var CountryEntity|null $country */
+        $country = $this->countryRepository->search($criteria, $context)->first();
+
+        if (!$country instanceof CountryEntity) {
+            return null;
+        }
+
+        return $country->getId();
+    }
+
+    public function getSalutationId(Context $context): string
+    {
+        $criteria = new Criteria();
+        $criteria->addFilter(
+            new EqualsFilter('salutationKey', 'not_specified')
+        );
+
+        /** @var SalutationEntity|null $salutation */
+        $salutation = $this->salutationRepository->search($criteria, $context)->first();
+
+        if ($salutation === null) {
+            throw new RuntimeException();
+        }
+
+        return $salutation->getId();
+    }
+
+    /**
+     * @param array<string, string|null> $addressData
+     */
+    public function isIdenticalAddress(CustomerAddressEntity $address, array $addressData): bool
+    {
+        foreach (self::ADDRESS_KEYS as $key) {
+            if ($address->get($key) !== ($addressData[$key] ?? null)) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    public function getCustomerAddressId(CustomerEntity $customer, array $addressData)
+    {
+        $matchingAddress = null;
+
+        $addresses = $customer->getAddresses();
+        if ($addresses !== null) {
+            foreach ($addresses as $address) {
+                if ($this->isIdenticalAddress($address, $addressData)) {
+                    $matchingAddress = $address;
+
+                    break;
+                }
+            }
+        }
+
+        return $matchingAddress === null ? Uuid::randomHex() : $matchingAddress->getId();
+    }
+
+    public function getSalesChannelCountryIso(SalesChannelContext $salesChannelContext): ?string
+    {
+        $salesChannelId = $salesChannelContext->getSalesChannelId();
+
+        $criteria = new Criteria([$salesChannelId]);
+        $criteria->addAssociation('country');
+
+        $salesChannel = $this->salesChannelRepository->search(
+            $criteria,
+            $salesChannelContext->getContext()
+        )->first();
+
+        if (!$salesChannel->getCountry()) {
+            return null;
+        }
+
+        return (string) $salesChannel->getCountry()->getIso();
+    }
+}