+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.6.4
+- Fast checkout button for PayPal added
+- Fast checkout button for iDEAL added
+
 # 1.6.3
 - Fixed install payment methods issue